nfc: ndef: enable encoding of long uri messages
The NDEF URI Message module now supports long URI payload encoding.
It is possible to encode URIs longer than 255 bytes.

+++ include/nfc/ndef/uri_msg.h
@@ -47,7 +47,7 @@ extern "C" {
  */
 int nfc_ndef_uri_msg_encode(enum nfc_ndef_uri_rec_id uri_id_code,
 			    uint8_t const *const uri_data,
-			    uint8_t uri_data_len,
+			    uint16_t uri_data_len,
 			    uint8_t *buf,
 			    uint32_t *len);
